services/ai-service/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
httpx
PyJWT
